requirements.txt
langchain
langchain-openai
openai
black
python-dotenv
langchain-agents-pip